requirements.txt
streamlit
pandas
snowflake-connector-python
openai==0.28
python-dotenv